config: Why reinvent the wheel when we have `self.option()`?
I got annoyed by the lack of space between the default `[n]` and my
answer when running the module config wizard. Then when I went to add in
the missing space, I noticed that this code was duplicating the method
defined just above it. Why the hell not use it?